Fix badge workflow & jupyter dependency
The previous commit's dynamic badge generation failed as my repository's workflow permissions for the GITHUB_TOKEN was read only. I updated it to Read and Write in the repository settings. Then, separated out the dynamic badge updating logic to a update-badge.yml file with write permissions which runs upon completion of the main ci.yml workflow which does not have write permissions. Made this distinction for stricter permission control for safety reasons. Besides this, added jupyter dependency to poetry so we do not have to pip install it within ci.yml for every push or pull request.